Minor Roles
On servers with a under 18s segeregration policy, like this cogs main prupose server, if you have a `Minor` role, it will put minors in that role INSTEAD of the Trainer role upon approval. Uses the optional command place, so you have to choose between Spoofer and Minor.

If a player is both, use the best practice is to use the spoofer optional and then change the roles yourself manually.